修复了裂缝bug,正在调试一个基本impact--BasicSmoothPlane

+++ Assets/Scripts/SE/TerrainImpacts/BasicRandomAdjust.cs
@@ -11,15 +11,15 @@ namespace SE.TerrainImpacts {
 
             long[] map = Data.Map;
 
-            long
-                RandomRange = System.Math.Max(Data.Region.x2 - Data.Region.x1, Data.Region.y2 - Data.Region.y1) / 5,
+            ulong
+                RandomRange = (ulong)System.Math.Max(Data.Region.x2 - Data.Region.x1, Data.Region.y2 - Data.Region.y1) / 5,
                 t = RandomRange / 2;
 
-            map[1] += Data.Seed[1].NextRandomNum(RandomRange) - t;
-			map[3] += Data.Seed[2].NextRandomNum(RandomRange) - t;
-			map[4] += Data.Seed[0].NextRandomNum(RandomRange) - t;
-			map[5] += Data.Seed[3].NextRandomNum(RandomRange) - t;
-			map[7] += Data.Seed[4].NextRandomNum(RandomRange) - t;
+            map[1] += (long)(Data.Seed[1].NextRandomNum(RandomRange) - t);
+			map[3] += (long)(Data.Seed[2].NextRandomNum(RandomRange) - t);
+			map[4] += (long)(Data.Seed[0].NextRandomNum(RandomRange) - t);
+			map[5] += (long)(Data.Seed[3].NextRandomNum(RandomRange) - t);
+			map[7] += (long)(Data.Seed[4].NextRandomNum(RandomRange) - t);
         }
 	}
 }
